perf: chainId changed to uint256

+++ contracts/ShambaFireConsumer.sol
@@ -14,7 +14,7 @@ contract ShambaFireConsumer is ChainlinkClient, ShambaChainSelector, ShambaWhite
     mapping(uint256 => string) private cids;
 
     /// @param chain_id    ETH Chain Id of the network on which the contract is getting deployed
-    constructor(uint64 chain_id) ShambaChainSelector(chain_id) {
+    constructor(uint256 chain_id) ShambaChainSelector(chain_id) {
         shambaChainSelector = new ShambaChainSelector(chain_id);
         setChainlinkToken(shambaChainSelector.linkTokenContractAddress());
         setChainlinkOracle(shambaChainSelector.operatorAddress());
